pyproject.toml
[project]
name = "pyswip-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "openai>=2.1.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.11.10",
    "pyswip>=0.3.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "when-changed>=0.3.0",
]
